+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/test/LiftTest.java
@@ -1,8 +1,6 @@
 package org.firstinspires.ftc.teamcode.auto.test;
 
 
-import static org.firstinspires.ftc.teamcode.mechanisms.Lift.*;
-
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.SequentialAction;
 import com.acmerobotics.roadrunner.SleepAction;
@@ -10,8 +8,7 @@ import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.mechanisms.Intake;
-import org.firstinspires.ftc.teamcode.mechanisms.Lift;
+import org.firstinspires.ftc.teamcode.mechanisms.v1.Lift;
 
 
 @Config
