+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/test/ScoreTest.java
@@ -9,10 +9,10 @@ import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.mechanisms.Arm;
-import org.firstinspires.ftc.teamcode.mechanisms.Intake;
-import org.firstinspires.ftc.teamcode.mechanisms.Lift;
-import org.firstinspires.ftc.teamcode.mechanisms.Wrist;
+import org.firstinspires.ftc.teamcode.mechanisms.v1.Arm;
+import org.firstinspires.ftc.teamcode.mechanisms.v1.Intake;
+import org.firstinspires.ftc.teamcode.mechanisms.v1.Lift;
+import org.firstinspires.ftc.teamcode.mechanisms.v1.Wrist;
 
 @Config
 @Autonomous(name = "ScoreTest", group = "Test")
