+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/v1/Arm.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.mechanisms;
+package org.firstinspires.ftc.teamcode.mechanisms.v1;
 
 import androidx.annotation.NonNull;
 
@@ -9,8 +9,6 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
-
-import java.security.PublicKey;
 
 @Config
 public class Arm {
