+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/v1/Wrist.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.mechanisms;
+package org.firstinspires.ftc.teamcode.mechanisms.v1;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.Action;
